upgrade/boot_loader: Move MockDecryptor to test_doubles.
